+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 * pool: increase `MAX_ADJ_RETRY_SEC` to 120 secs ([Yuki Kishimoto])
 * pool: return reference instead of cloned structs for some getter methods of `Relay` ([Yuki Kishimoto])
 * pool: removed unnecessary timeout during the shutdown notification process ([Yuki Kishimoto])
+* pool: deprecate `RelaySendOptions::skip_disconnected` ([Yuki Kishimoto])
 * sdk: deprecate `Client::get_events_of` and `Client::get_events_from` methods ([Yuki Kishimoto])
 * sdk: use `Events` instead of `Vec<Event>` in fetch and query methods ([Yuki Kishimoto])
 * sdk: rename `stream_events_of` to `stream_events` ([Yuki Kishimoto])